backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
python-dotenv==1.0.1
motor==3.6.0
pydantic==2.9.2
python-multipart==0.0.12
httpx==0.27.2